+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Update docs example in travelingsalesman.
 - Update the documentation so that the examples are runnable by copy-pasting.
+- Add example in the docs of game solver.
 
 ### 0.2.9